# api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: dsgamer777@gmail.com
    name: Omado
    url: https://twitter.com/shallow_omado
  description: アカウント関連API
  license:
    name: GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.ja.html
  title: UsagiBooru Accounts API
  version: "2.0"
servers:
- description: Prod
  url: https://api.gochiusa.team
- description: Staging
  url: https://api-next.gochiusa.team
- description: Dev
  url: http://127.0.0.1:8080
security:
- Authorization: []
...
paths:
...
  /accounts/{accountID}/timeline/follow:
    post:
      description: 絵師をフォローします
      operationId: followArtist
      parameters:
      - description: 対象のアカウントID
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LightArtistStruct'
      responses:
        "204":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: No Content
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Conflict
      summary: Follow artist
      tags:
      - timeline
  /accounts/{accountID}/timeline/unfollow:
    post:
      description: 絵師のフォローを解除します
      operationId: unfollowArtist
      parameters:
      - description: 対象のアカウントID
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LightArtistStruct'
      responses:
        "204":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: No Content
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Conflict
      summary: Unfollow artist
      tags:
      - timeline
  /accounts/{accountID}/timeline:
    get:
      description: フォロー中の絵師ID一覧を取得します
      operationId: getFollowingArtists
      parameters:
      - description: 対象のアカウントID
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          type: integer
        style: simple
      - description: ソート方法 l:いいね数順 d:投稿日順
        explode: true
        in: query
        name: sort
        required: false
        schema:
          default: d
          enum:
          - l
          - d
          example: d
          type: string
        style: form
      - description: ソート方向 d:降順(大→小) a:昇順(小→大)
        explode: true
        in: query
        name: order
        required: false
        schema:
          default: d
          enum:
          - d
          - a
          example: d
          maxLength: 1
          minLength: 1
          type: string
        style: form
      - description: 'ページ番号 (NOTE: 1ページ辺りの応答は20件)'
        explode: true
        in: query
        name: page
        required: false
        schema:
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTimelineFollowingResponse'
          description: OK
      summary: Get timeline followings
      tags:
      - timeline
...
components:
...
  schemas:
...
    GeneralMessageResponse:
      example:
        message: message
      properties:
        message:
          description: 応答メッセージ(何が間違っているかの詳細)
          minLength: 1
          type: string
      required:
      - message
      title: GeneralMessageResponse
      type: object
      x-examples:
        not-found:
          message: Specified content was not found.
        no-permission:
          message: You don't have enough permission to do it.
        bad-request:
          message: Your request body is wrong.
        locked:
          message: Specified content is referenced from other content.
...
    LightArtistStruct:
      example:
        name: 彩電
        artistID: 1
      properties:
        artistID:
          example: 1
          type: integer
        name:
          example: 彩電
          type: string
      title: LightArtistStruct
      type: object
      x-examples:
        example:
          artistID: 1
          name: 彩電
    GetTimelineFollowingResponse:
      example:
        pagination:
          current: 1
          pages: 1
          perPage: 20
          count: 1
          title: 香風智乃
          type: tag
        follows:
        - name: 彩電
          artistID: 1
        - name: 彩電
          artistID: 1
      properties:
        pagination:
          $ref: '#/components/schemas/PaginationStruct'
        follows:
          items:
            $ref: '#/components/schemas/LightArtistStruct'
          type: array
      title: GetTimelineFollowingResponse
      type: object
      x-examples:
        example:
          pagination:
            title: フォロー一覧
            type: follow
            count: 1
            current: 1
            pages: 1
            perPage: 20
          follows:
          - artistID: 1
            name: 彩電
...
    PaginationStruct:
      example:
        current: 1
        pages: 1
        perPage: 20
        count: 1
        title: 香風智乃
        type: tag
      properties:
        title:
          description: ページネーションタイトル(表示用)
          example: 香風智乃
          minLength: 1
          type: string
        type:
          default: tag
          description: ページネーション種別(表示用)
          enum:
          - artist
          - tag
          - uploader
          - notify-client
          - notify-condition
          - mute
          - follow
          - article
          - mylist-list
          - art
          - upload-history
          example: tag
          type: string
        count:
          default: 0
          description: ヒット総数
          example: 1
          minimum: 0
          type: integer
        current:
          default: 1
          description: 現在のページ
          example: 1
          minimum: 1
          type: integer
        pages:
          default: 1
          description: ページ数
          example: 1
          minimum: 1
          type: integer
        perPage:
          default: 20
          description: ページ毎の取得数
          enum:
          - 20
          - 40
          - 60
          - 80
          - 100
          example: 20
          format: int32
          maximum: 100
          minimum: 20
          type: integer
      required:
      - count
      - current
      - pages
      - perPage
      - title
      - type
      title: PaginationStruct
      type: object
      x-examples:
        example:
          title: 香風智乃
          type: tag
          count: 1
          current: 1
          pages: 1
          perPage: 20
...
  securitySchemes:
    Authorization:
      scheme: bearer
      type: http